Prototype a Ledger WebUSB bridge entry point
The endpoint can be built with the extension, but needs to be delivered
in a secure context (from an SSL-protected site in a new tab).

Using this endpoint and a new keyring, the hope is to support Ledger via
WebUSB on Chrome and U2F on Firefox, skipping the "connect to
LedgerLive" hack some other wallets have gone with.

Refs #336 #337