KeyringService encrypts vault data without stringifying
The encryption code stringifies, so the previuos version stringified
twice, resulting in decrypted data that was still in string form. The
code that interacted with the decrypted vaults expected the decrypted
data to already be parsed JSON, rather than still-stringified JSON. This
adjustment stringifies just once, ensuring the read code gets a decoded
JSON object.