Persist keyrings after adding addresses
The way the code was written, new keyrings were added and persisted,
then new addresses were added. The result of this sequencing was that
the persisted keyring did not correctly track how many addresses had
been generated, leading to a failure to track those addresses after
locking and then unlocking the vault (or any other sequence of actions
that led to restoring the vaults from serialized storage).